create action menu for nav inputs with menus
